Add handling for when a stream is closed and then errors
One of the error cases with legacy streams is that the stream can be
closed via an error. In such cases, it is common for the underlying
stream to emit an 'error' event, followed by a 'close' event. In this
situation, stream worker should listen to the close event and consider
that closing the stream as well.

https://github.com/nodejs/node-v0.x-archive/blob/master/doc/api/stream.markdown#event-close